clean up

+++ include/helper.h
@@ -13,8 +13,8 @@ void createScatterPlot(std::vector<double> x, std::vector<double> y, std::string
 
 void createHistPlot(std::vector<double> y, double bins = 10, std::string xlabel = "", std::string ylabel = "",
                     std::string title = "");
-void createBarPlot(std::vector<double> x, std::vector<double> y, std::string xlabel = "", std::string ylabel = "",
-                   std::string title = "");
+void createBarPlot(std::vector<long double> x, std::vector<long double> y, std::string xlabel = "",
+                   std::string ylabel = "", std::string title = "");
 void createMultiPlot(std::vector<std::vector<double>> vectorX, std::vector<std::vector<double>> y,
                      std::vector<std::string> xlabel, std::vector<std::string> ylabel, std::vector<std::string> title);
 double transformationMethodPoisson(double lambda, std::optional<double> randomValue);
